% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cwlParam.R
\name{baseCommand}
\alias{baseCommand}
\alias{baseCommand<-}
\title{baseCommand}
\usage{
baseCommand(object)

baseCommand(cwl) <- value
}
\arguments{
\item{cwl}{A cwlParam object}

\item{value}{Specifies the program to execute.}
}
\description{
baseCommand

baseCommand
}
