% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cwlParam.R
\name{cwlVersion}
\alias{cwlVersion}
\alias{cwlClass}
\alias{cwlClass<-}
\alias{baseCommand}
\title{cwlVersion}
\usage{
cwlVersion(cwl)

cwlClass(cwl)

cwlClass(cwl) <- value

baseCommand(cwl)
}
\arguments{
\item{cwl}{A cwlParam object}

\item{value}{The cwlParam class.}
}
\description{
cwlVersion

cwlClass

cwlClass

baseCommand
}
