% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cwlParam.R, R/stepParam.R
\name{cwlVersion}
\alias{cwlVersion}
\alias{cwlVersion<-}
\alias{cwlClass}
\alias{cwlClass<-}
\alias{baseCommand}
\alias{baseCommand<-}
\alias{arguments}
\alias{arguments<-}
\alias{hints}
\alias{hints<-}
\alias{requirements}
\alias{requirements<-}
\alias{stdOut}
\alias{stdOut<-}
\alias{short}
\alias{runs}
\alias{steps}
\alias{steps<-}
\title{cwlParam methods}
\usage{
cwlVersion(cwl)

cwlVersion(cwl) <- value

cwlClass(cwl)

cwlClass(cwl) <- value

baseCommand(cwl)

baseCommand(cwl) <- value

arguments(cwl)

arguments(cwl) <- value

hints(cwl)

hints(cwl) <- value

requirements(cwl)

requirements(cwl) <- value

stdOut(cwl)

stdOut(cwl) <- value

short(object)

runs(object)

steps(cwl)

steps(cwl) <- value
}
\arguments{
\item{cwl}{A `cwlParam` object.}

\item{value}{Assign value to the `cwlParam` object.}

\item{object}{An cwlParam or cwlStepParam object}
}
\description{
The function to show short summary of cwlParam or cwlStepParam

The function to access all runs of a cwlStepParam object
}
