% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllClasses.R, R/cwlParam.R
\docType{class}
\name{cwlParam-class}
\alias{cwlParam-class}
\alias{cwlParam}
\alias{$,cwlParam-method}
\alias{$<-,cwlParam-method}
\title{Parameters for CWL}
\usage{
cwlParam(cwlVersion = "v1.0", cwlClass = "CommandLineTool",
  baseCommand = character(), requirements = list(), hints = list(),
  arguments = list(), inputs = InputParamList(),
  outputs = OutputParamList(), stdout = character())

\S4method{$}{cwlParam}(x, name)

\S4method{$}{cwlParam}(x, name) <- value
}
\arguments{
\item{cwlVersion}{CWL version}

\item{cwlClass}{"CommandLineTool"}

\item{baseCommand}{Specifies the program to execute}

\item{requirements}{Requirements that apply to either the runtime environment or the workflow engine.}

\item{hints}{Any or a list for the workflow engine.}

\item{arguments}{Command line bindings which are not directly associated with input parameters.}

\item{inputs}{A object of `InputParamList`.}

\item{outputs}{A object of `OutputParamList`.}

\item{stdout}{Capture the command's standard output stream to a file written to the designated output directory.}

\item{name}{One name of input list}

\item{value}{Set the value to the input list by name}

\item{name}{One one of input list}
}
\description{
Parameters for CWL

Parameters for CWL
The main CWL parameter class and constructor for command tools. More details: https://www.commonwl.org/v1.0/CommandLineTool.html

Extract input values by name

Set input values by name
}
