% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cwlShiny.R
\name{cwlShiny}
\alias{cwlShiny}
\title{cwlShiny}
\usage{
cwlShiny(cwl, inputList = list(), upload = FALSE, ...)
}
\arguments{
\item{cwl}{A cwlParam object.}

\item{inputList}{a list of choices for the inputs of cwl
object. The name of the list must match the inputs of the cwl
object.}

\item{upload}{Whether to upload file. If FALSE, the upload field
will be text input (file path) instead of file input.}

\item{...}{More options for `runCWL`.}
}
\description{
Function to generate shiny app automaticlly for a `cwlParam`
object.
}
\examples{
input1 <- InputParam(id = "sth")
echo <- cwlParam(baseCommand = "echo", inputs = InputParamList(input1))
echoApp <- cwlShiny(echo)
}
