% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllClasses.R
\docType{class}
\name{cwlStepList-class}
\alias{cwlStepList-class}
\alias{cwlStepList}
\title{cwlStepList}
\usage{
cwlStepList(...)
}
\arguments{
\item{...}{A list of `cwlStep`.}
}
\value{
An object of class `cwlStepList`.
}
\description{
cwlStepList

cwlStepList
}
\examples{
s1 <- cwlStep(id = "s1")
cwlStepList(s1)
}
