% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllClasses.R
\docType{class}
\name{cwlStepParam-class}
\alias{cwlStepParam-class}
\alias{cwlStepParam}
\title{cwlStepParam}
\usage{
cwlStepParam(cwlVersion = "v1.0", cwlClass = "Workflow",
  requirements = list(), id = character(), hints = list(),
  arguments = list(), inputs = InputParamList(),
  outputs = OutputParamList(), stdout = character(),
  steps = stepParamList())
}
\arguments{
\item{cwlVersion}{CWL version}

\item{cwlClass}{"Workflow".}

\item{requirements}{Requirements that apply to either the runtime
environment or the workflow engine.}

\item{id}{The unique identifier for this process object.}

\item{hints}{Any or a list for the workflow engine.}

\item{arguments}{Command line bindings which are not directly
associated with input parameters.}

\item{inputs}{A object of `InputParamList`.}

\item{outputs}{A object of `OutputParamList`.}

\item{stdout}{Capture the command's standard output stream to a
file written to the designated output directory.}

\item{steps}{A list of `stepParamList`.}
}
\description{
A workflow steps paramter, which connect multiple command line
steps into a workflow. More details: stepInParamList.
}
