% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cwlParam.R
\name{cwlVersion<-}
\alias{cwlVersion<-}
\title{cwlVersion
CWL document version}
\usage{
cwlVersion(cwl) <- value
}
\arguments{
\item{cwl}{A cwlParam object.}

\item{value}{CWL document version.}
}
\description{
cwlVersion
CWL document version
}
