% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cwlProcessWorkflow-methods.R
\name{cwlWorkflow-methods}
\alias{cwlWorkflow-methods}
\alias{runs}
\title{cwlWorkflow methods}
\usage{
runs(object)
}
\arguments{
\item{object}{A `cwlWorkflow` object.}
}
\value{
`cwlProcess` objects or paths of CWL file.
}
\description{
runs: The function to access all runs of a `cwlWorkflow` object.
}
\examples{
s1 <- cwlWorkflow()
runs(s1)
s1
short(s1)
}
