% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/basilisk.R
\docType{data}
\name{env_Rcwl}
\alias{env_Rcwl}
\title{Rcwl conda environment}
\format{
An object of class \code{BasiliskEnvironment} of length 1.
}
\usage{
env_Rcwl
}
\description{
Rcwl conda envrionment to install `cwltool` by basilisk.
}
\keyword{datasets}
