% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllClasses.R
\docType{class}
\name{InputParam-class}
\alias{InputParam-class}
\alias{InputParam}
\title{InputParam}
\usage{
InputParam(id, label = "", type = "string", position = 0L,
  prefix = "", separate = TRUE, itemSeparator = character(),
  valueFrom = character(), default = character(),
  value = character())
}
\arguments{
\item{id}{The unique identifier for this parameter object.}

\item{label}{A short, human-readable label of this object.}

\item{type}{valid types of data that may be assigned to this parameter.}

\item{position}{The position for this parameter.}

\item{prefix}{Command line prefix to add before the value.}

\item{separate}{If true (default), then the prefix and value must be added as separate command line arguments; if false, prefix and value must be concatenated into a single command line argument.}

\item{itemSeparator}{Join the array elements into a single string with the elements separated by by itemSeparator.}

\item{valueFrom}{String or Expression.}

\item{default}{The default value for this parameter}

\item{value}{Assigned value for this parameter}
}
\description{
InputParam

InputParam
parameter for a command tool. More details: https://www.commonwl.org/v1.0/CommandLineTool.html#CommandInputParameter
}
