% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllClasses.R, R/cwlParam.R
\docType{class}
\name{InputParamList-class}
\alias{InputParamList-class}
\alias{InputParamList}
\alias{inputs}
\title{InputParamList}
\usage{
InputParamList(...)

inputs(cwl)
}
\arguments{
\item{...}{The InputParam objects.}

\item{cwl}{A cwlParam object}
}
\value{
An object of class `InputParamList`.

inputs: A list of `InputParam`.
}
\description{
InputParamList

InputParamList
A list of InputParam

inputs
}
\examples{
input1 <- InputParam(id = "sth")
InputParamList(input1)
## Inputs
input1 <- InputParam(id = "sth")
echo <- cwlParam(baseCommand = "echo", inputs = InputParamList(input1))
inputs(echo)
}
