% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllClasses.R
\docType{class}
\name{OutputArrayParam-class}
\alias{OutputArrayParam-class}
\alias{OutputArrayParam}
\title{Output array parameters}
\usage{
OutputArrayParam(label = character(), type = "array",
  items = character(), glob = character(), loadContents = logical(),
  outputEval = character())
}
\arguments{
\item{label}{A short, human-readable label of this object.}

\item{type}{Must be "array".}

\item{items}{Defines the type of the array elements.}

\item{glob}{Pattern to find files relative to the output directory.}

\item{loadContents}{Read text from globbed file.}

\item{outputEval}{Evaluate an expression to generate the output
value.}
}
\description{
Parameters for array outputs. More details:
https://www.commonwl.org/v1.0/CommandLineTool.html#CommandOutputArraySchema
}
