% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotCWL.R
\name{plotCWL}
\alias{plotCWL}
\title{plotCWL}
\usage{
plotCWL(cwl, ...)
}
\arguments{
\item{cwl}{A cwlStepParam object to plot}

\item{...}{other parameters from `mermaid` function}
}
\description{
Function to plot cwlStepParam object.
}
