% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stepParam.R
\docType{methods}
\name{+,cwlStepParam,stepParam-method}
\alias{+,cwlStepParam,stepParam-method}
\title{Pipeline}
\usage{
\S4method{+}{cwlStepParam,stepParam}(e1, e2)
}
\arguments{
\item{e1}{A `cwlStepParam` object.}

\item{e2}{A `stepParam` object.}
}
\description{
To build a pipeline by connecting multiple `stepParam` to a
`cwlStepParam` object.
}
