% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stepParam.R
\name{+,cwlWorkflow,stepParam-method}
\alias{+,cwlWorkflow,stepParam-method}
\title{Pipeline}
\usage{
\S4method{+}{cwlWorkflow,stepParam}(e1, e2)
}
\arguments{
\item{e1}{A `cwlWorkflow` object.}

\item{e2}{A `stepParam` object.}
}
\value{
A `cwlWorkflow` object.
}
\description{
To build a pipeline by connecting multiple `stepParam` to a
`cwlWorkflow` object.
}
\seealso{
\code{\link{cwlWorkflow}}
}
