% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Rcwl-package.R
\docType{package}
\name{Rcwl-package}
\alias{Rcwl-package}
\alias{Rcwl}
\alias{Rcwl,Rcwl-package}
\title{Rcwl-package}
\description{
An R package to wrap command line tools and build pipelines with
Common Workflow Language (CWL).
}
\seealso{
\code{\link{cwlProcess}}

\code{\link{cwlWorkflow}}

\code{\link{cwlStep}}

\code{\link{runCWL}}
}
