% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Rcwl-package.R
\docType{package}
\name{Rcwl}
\alias{Rcwl}
\title{Rcwl}
\description{
An R package to wrap command line tools and build pipelines with
Common Workflow Language.
}
\seealso{
\code{\link{cwlParam}}

\code{\link{runCWL}}
}
