% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cwlParam.R, R/stepParam.R
\name{requireDocker}
\alias{requireDocker}
\alias{requireJS}
\alias{requireSoftware}
\alias{requireInitialWorkDir}
\alias{requireSubworkflow}
\alias{requireScatter}
\alias{requireMultipleInput}
\alias{requireStepInputExpression}
\title{requireDocker}
\usage{
requireDocker(
  docker = NULL,
  Load = NULL,
  File = NULL,
  Import = NULL,
  ImageId = NULL,
  OutputDir = NULL
)

requireJS(expressionLib = list())

requireSoftware(packages = list())

requireInitialWorkDir(listing = list())

requireSubworkflow()

requireScatter()

requireMultipleInput()

requireStepInputExpression()
}
\arguments{
\item{docker}{The docker pull address.}

\item{Load}{dockerLoad}

\item{File}{dockerFile}

\item{Import}{dockerImport}

\item{ImageId}{dockerImageId}

\item{OutputDir}{dockerOutputDirectory}

\item{expressionLib}{optional code}

\item{packages}{The list of software to be configured.}

\item{listing}{The list of files or directories.}
}
\value{
A DockerRequirement list

A InlineJavascriptRequirement list

A SoftwareRequirement list

A InitialWorkDirRequirement list

A SubworkflowFeatureRequirement list

A ScatterFeatureRequirement list

A MultipleInputFeatureRequirement list

A StepInputExpressionRequirement list
}
\description{
requireDocker

requireJS

requireSoftware

InitialWorkDirRequirement

SubworkflowFeatureRequirement

ScatterFeatureRequirement

MultipleInputFeatureRequirement

StepInputExpressionRequirement
}
