% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/runCWL.R
\name{runCWL}
\alias{runCWL}
\title{run cwlParam
Execute a cwlParam object with assigned inputs.}
\usage{
runCWL(cwl, prefix = tempfile(), cwlRunner = "cwltool",
  Args = character(), stdout = TRUE, stderr = TRUE, ...)
}
\arguments{
\item{cwl}{A `cwlParam` or `cwlStepParam` object.}

\item{prefix}{The prefix of `cwl` and `yml` file to write.}

\item{cwlRunner}{The path to the `cwltool` or `cwl-runner`.}

\item{Args}{The arguments for `cwltool` or `cwl-runner`. For example, "--debug" can work with `cwltool` to show debug information.}

\item{stdout}{standard output from `system2`.}

\item{stderr}{standard error from `system2`.}

\item{...}{The other options from `writeCWL` and `system2`.}
}
\description{
run cwlParam
Execute a cwlParam object with assigned inputs.
}
