% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/runCWL.R
\name{runCWLBatch}
\alias{runCWLBatch}
\title{run CWL with batchtools}
\usage{
runCWLBatch(cwl, wdir = getwd(), inputList, paramList = list(),
  BPoptions = list(), ...)
}
\arguments{
\item{cwl}{A `cwlParam` or `cwlStepParam` object.}

\item{wdir}{Directory to output results}

\item{inputList}{An input list to run in parallel. The list names must be in the inputs of cwl. Jobs will be submitted in parallel for each element in the list. The output directory of each job will be made using the name of each element under the `wdir`.}

\item{paramList}{A parameter list for the cwl. The list names must be in the inputs of cwl.}

\item{BPoptions}{The options for `BatchtoolsParam`.}

\item{...}{The options from runCWL.}
}
\description{
run CWL with batchtools
}
