% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/runCWL.R
\name{runCWLBP}
\alias{runCWLBP}
\title{run CWL with BiocParallel}
\usage{
runCWLBP(cwl, outdir, BPPARAM, ...)
}
\arguments{
\item{cwl}{cwl A `cwlProcess` or `cwlWorkflow` object.}

\item{outdir}{Directory for output results}

\item{BPPARAM}{The options for `BiocParallelParam`.}

\item{...}{The other options from runCWL.}
}
\value{
Results from computing nodes and logs from cwltool.
}
\description{
Submit one CWL object with assigned values with BiocParallel.
}
