% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cwlProcess.R
\name{runs}
\alias{runs}
\title{runs}
\usage{
runs(object)
}
\arguments{
\item{object}{A cwlWorkflow object.}
}
\value{
cwlProcess objects or paths of CWL file.
}
\description{
The function to access all runs of a cwlWorkflow object
}
\examples{
s1 <- cwlWorkflow()
runs(s1)
}
