% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cwlProcess.R
\name{short}
\alias{short}
\title{short}
\usage{
short(object)
}
\arguments{
\item{object}{An cwlProcess or cwlWorkflow object}
}
\value{
A short summary of an object of cwlProcess or cwlWorkflow.
}
\description{
The function to show short summary of cwlProcess or cwlWorkflow
}
\examples{
s1 <- cwlWorkflow()
short(s1)
}
