% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stepParam.R
\name{Step}
\alias{Step}
\title{Step function
Function to assign value to `stepParam` object.}
\usage{
Step(id, run = cwlParam(), In = list())
}
\arguments{
\item{id}{The id of `stepParam` object.}

\item{run}{A `cwlParam` object for command tool.}

\item{In}{one or two layes of list.}
}
\description{
Step function
Function to assign value to `stepParam` object.
}
