% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllClasses.R
\docType{class}
\name{stepInParamList-class}
\alias{stepInParamList-class}
\alias{stepInParamList}
\title{stepInParamList}
\usage{
stepInParamList(...)
}
\arguments{
\item{...}{A list of `stepInParam` objects.}
}
\description{
stepInParamList

stepInParamList
}
