% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cwlStep.R
\name{stepInputs}
\alias{stepInputs}
\title{stepInputs}
\usage{
stepInputs(stepList)
}
\arguments{
\item{stepList}{a list of `cwlStep` objects.}
}
\value{
InputParamList.
}
\description{
prepare inputs for workflow from `cwlStep` objects
}
