% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cwlStep.R
\name{stepOutputs}
\alias{stepOutputs}
\title{stepOutputs}
\usage{
stepOutputs(stepList)
}
\arguments{
\item{stepList}{a list of `cwlStep` objects.}
}
\value{
OutputParamList.
}
\description{
prepare outputs for workflow from `cwlStep` objects
}
