% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllClasses.R
\docType{class}
\name{stepParamList-class}
\alias{stepParamList-class}
\alias{stepParamList}
\title{stepParamList}
\usage{
stepParamList(...)
}
\arguments{
\item{...}{A list of `stepParam`.}
}
\value{
An object of class `stepParamList`.
}
\description{
stepParamList

stepParamList
}
\examples{
s1 <- stepParam(id = "s1")
stepParamList(s1)
}
