% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stepParam.R
\name{steps}
\alias{steps}
\alias{steps<-}
\title{Steps}
\usage{
steps(cwl)

steps(cwl) <- value
}
\arguments{
\item{cwl}{A cwlWorkflow object.}

\item{value}{A list of steps.}
}
\value{
steps: A list of stepParam objects.
}
\description{
Function to extract step slots
}
\seealso{
\code{\link{cwlWorkflow}}
}
