% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cwlParam.R
\name{cwlClass}
\alias{cwlClass}
\alias{cwlClass<-}
\title{cwlClass}
\usage{
cwlClass(cwl)

cwlClass(cwl) <- value
}
\arguments{
\item{cwl}{A cwlParam object}

\item{value}{The cwlParam class.}
}
\description{
cwlClass

cwlClass
}
