% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cwlParam.R
\name{cwlVersion}
\alias{cwlVersion}
\alias{cwlVersion<-}
\alias{cwlClass}
\alias{cwlClass<-}
\alias{baseCommand}
\alias{baseCommand<-}
\alias{arguments}
\alias{arguments<-}
\alias{hints}
\alias{hints<-}
\alias{requirements}
\alias{requirements<-}
\title{cwlParam methods}
\usage{
cwlVersion(cwl)

cwlVersion(cwl) <- value

cwlClass(cwl)

cwlClass(cwl) <- value

baseCommand(cwl)

baseCommand(cwl) <- value

arguments(cwl)

arguments(cwl) <- value

hints(cwl)

hints(cwl) <- value

requirements(cwl)

requirements(cwl) <- value
}
\arguments{
\item{cwl}{A `cwlParam` object.}

\item{value}{Assign value to the `cwlParam` object.}
}
\description{
cwlParam methods

cwlVersion
CWL document version

cwlClass

cwlClass

baseCommand

baseCommand

arguments

arguments

hints

hints

requirements

requirements
}
