% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllClasses.R
\docType{class}
\name{cwlParam-class}
\alias{cwlParam-class}
\alias{cwlParam}
\title{Parameters for CWL}
\usage{
cwlParam(cwlVersion = "v1.0", cwlClass = "CommandLineTool",
  baseCommand = character(), requirements = list(), hints = list(),
  arguments = list(), id = character(), label = character(),
  inputs = InputParamList(), outputs = OutputParamList(),
  stdout = character())
}
\arguments{
\item{cwlVersion}{CWL version}

\item{cwlClass}{"CommandLineTool"}

\item{baseCommand}{Specifies the program to execute}

\item{requirements}{A list of Requirement lists that apply to
either the runtime environment or the workflow engine.}

\item{hints}{Any or a list for the workflow engine.}

\item{arguments}{Command line bindings which are not directly
associated with input parameters.}

\item{id}{The unique identifier for this process object.}

\item{label}{A short, human-readable label of this process object.}

\item{inputs}{A object of `InputParamList`.}

\item{outputs}{A object of `OutputParamList`.}

\item{stdout}{Capture the command's standard output stream to a
file written to the designated output directory.}
}
\description{
The main CWL parameter class and constructor for command
tools. More details:
https://www.commonwl.org/v1.0/CommandLineTool.html
}
\details{
https://www.commonwl.org/v1.0/CommandLineTool.html
}
\examples{
input1 <- InputParam(id = "sth")
echo <- cwlParam(baseCommand = "echo", inputs = InputParamList(input1))
echo$sth <- "Hello World!"
r1 <- runCWL(echo)
}
