% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cwlParam.R
\name{cwlVersion}
\alias{cwlVersion}
\alias{cwlVersion<-}
\title{cwlVersion}
\usage{
cwlVersion(cwl)

cwlVersion(cwl) <- value
}
\arguments{
\item{cwl}{A cwlParam object.}

\item{value}{CWL document version.}
}
\description{
cwlVersion

cwlVersion
CWL document version
}
