% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllClasses.R, R/cwlParam.R
\docType{class}
\name{InputParam-class}
\alias{InputParam-class}
\alias{InputParam}
\alias{$,cwlParam-method}
\alias{$<-,cwlParam-method}
\title{Input parameters
InputParam}
\usage{
InputParam(
  id,
  label = "",
  type = "string",
  doc = character(),
  secondaryFiles = character(),
  streamable = logical(),
  format = character(),
  loadListing = character(),
  loadContents = logical(),
  position = 0L,
  prefix = "",
  separate = TRUE,
  itemSeparator = character(),
  valueFrom = character(),
  shellQuote = logical(),
  default = character(),
  value = character()
)

\S4method{$}{cwlParam}(x, name)

\S4method{$}{cwlParam}(x, name) <- value
}
\arguments{
\item{id}{The unique identifier for this parameter object.}

\item{label}{A short, human-readable label of this object.}

\item{type}{valid types of data that may be assigned to this
parameter.}

\item{doc}{Optional. This argument takes an arbitrary documentation
as a note for this object.}

\item{secondaryFiles}{Only valid when type: File or is an array of
items: File. Provides a pattern or expression specifying files
or directories that must be included alongside the primary
file.}

\item{streamable}{Only valid when type: File or is an array of
items: File. A value of true indicates that the file is read or
written sequentially without seeking.}

\item{format}{Only valid when type: File or is an array of items:
File.}

\item{loadListing}{Only valid when type: Directory or is an array
of items: Directory.}

\item{loadContents}{Only valid when type: File or is an array of
items: File.}

\item{position}{The position for this parameter.}

\item{prefix}{Command line prefix to add before the value.}

\item{separate}{If true (default), then the prefix and value must
be added as separate command line arguments; if false, prefix
and value must be concatenated into a single command line
argument.}

\item{itemSeparator}{Join the array elements into a single string
with the elements separated by by itemSeparator.}

\item{valueFrom}{String or Expression.}

\item{shellQuote}{If ShellCommandRequirement is in the requirements
for the current command, this controls whether the value is
quoted on the command line (default is true).}

\item{default}{The default value for this parameter}

\item{value}{Assigned value for this parameter}

\item{x}{A `cwlParam` object.}

\item{name}{One one of input list}
}
\value{
An object of class `InputParam`.
}
\description{
parameter for a command tool. More details:
https://www.commonwl.org/v1.0/CommandLineTool.html#CommandInputParameter
}
\examples{
input1 <- InputParam(id = "sth")
}
