% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllClasses.R, R/cwlParam.R
\docType{class}
\name{InputParam-class}
\alias{InputParam-class}
\alias{InputParam}
\alias{$,cwlParam-method}
\alias{$<-,cwlParam-method}
\title{InputParam}
\usage{
InputParam(id, label = "", type = "string", doc = character(),
  secondaryFiles = character(), position = 0L, prefix = "",
  separate = TRUE, itemSeparator = character(),
  valueFrom = character(), default = character(),
  value = character())

\S4method{$}{cwlParam}(x, name)

\S4method{$}{cwlParam}(x, name) <- value
}
\arguments{
\item{id}{The unique identifier for this parameter object.}

\item{label}{A short, human-readable label of this object.}

\item{type}{valid types of data that may be assigned to this parameter.}

\item{doc}{A documentation string for this type.}

\item{secondaryFiles}{Only valid when type: File or is an array of items: File. Provides a pattern or expression specifying files or directories that must be included alongside the primary file.}

\item{position}{The position for this parameter.}

\item{prefix}{Command line prefix to add before the value.}

\item{separate}{If true (default), then the prefix and value must be added as separate command line arguments; if false, prefix and value must be concatenated into a single command line argument.}

\item{itemSeparator}{Join the array elements into a single string with the elements separated by by itemSeparator.}

\item{valueFrom}{String or Expression.}

\item{default}{The default value for this parameter}

\item{value}{Assigned value for this parameter}

\item{x}{A `cwlParam` object.}

\item{name}{One one of input list}
}
\description{
InputParam

InputParam
parameter for a command tool. More details: https://www.commonwl.org/v1.0/CommandLineTool.html#CommandInputParameter

Extract input values by name

Set input values by name
}
