% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllClasses.R
\docType{class}
\name{InputParamList-class}
\alias{InputParamList-class}
\alias{InputParamList}
\title{InputParamList}
\usage{
InputParamList(...)
}
\arguments{
\item{...}{The InputParam objects.}
}
\description{
InputParamList

InputParamList
A list of InputParam
}
