% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllClasses.R
\docType{class}
\name{OutputParam-class}
\alias{OutputParam-class}
\alias{OutputParam}
\title{Output parameters}
\usage{
OutputParam(id = "output", label = character(), type = "stdout",
  secondaryFiles = character(), streamable = logical(),
  glob = character(), loadContents = logical(),
  outputEval = character(), outputSource = character())
}
\arguments{
\item{id}{The unique identifier for this parameter object.}

\item{label}{A short, human-readable label of this object.}

\item{type}{Specify valid types of data that may be assigned to
this parameter.}

\item{secondaryFiles}{Provides a pattern or expression specifying
files or directories. Only valid when type: File or is an array
of items: File.}

\item{streamable}{A value of true indicates that the file is read
or written sequentially without seeking. Only valid when type:
File or is an array of items: File.}

\item{glob}{Pattern to find files relative to the output directory.}

\item{loadContents}{Read text from globbed file.}

\item{outputEval}{Evaluate an expression to generate the output
value.}

\item{outputSource}{Specifies one or more workflow parameters that
supply the value of to the output parameter.}
}
\value{
An object of class `OutputParam`.
}
\description{
An output parameter for a Command Line Tool. More details:
https://www.commonwl.org/v1.0/CommandLineTool.html#CommandOutputParameter
}
\examples{
o1 <- OutputParam(id = "file", type = "File", glob = "*.txt")
}
