% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllClasses.R, R/cwlParam.R
\docType{class}
\name{OutputParamList-class}
\alias{OutputParamList-class}
\alias{OutputParamList}
\alias{outputs}
\title{OutputParamList}
\usage{
OutputParamList(out = OutputParam(), ...)

outputs(cwl)
}
\arguments{
\item{out}{The default stdout parameter.}

\item{...}{The InputParam objects.}

\item{cwl}{A cwlParam object}
}
\description{
OutputParamList

OutputParamList
#' A list of InputParam

outputs
The outputs of a cwlParam object
}
