% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllClasses.R
\docType{class}
\name{OutputParamList-class}
\alias{OutputParamList-class}
\alias{OutputParamList}
\title{OutputParamList}
\usage{
OutputParamList(out = OutputParam(), ...)
}
\arguments{
\item{out}{The default stdout parameter.}

\item{...}{The InputParam objects.}
}
\description{
OutputParamList

OutputParamList
#' A list of InputParam
}
