% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cwlProcess.R, R/requirements.R
\name{requireDocker}
\alias{requireDocker}
\alias{requireJS}
\alias{requireSoftware}
\alias{requireInitialWorkDir}
\alias{Dirent}
\alias{requireManifest}
\alias{requireSubworkflow}
\alias{requireScatter}
\alias{requireMultipleInput}
\alias{requireStepInputExpression}
\title{requireDocker}
\usage{
requireDocker(
  docker = NULL,
  Load = NULL,
  File = NULL,
  Import = NULL,
  ImageId = NULL,
  OutputDir = NULL
)

requireJS(expressionLib = list())

requireSoftware(packages = list())

requireInitialWorkDir(listing = list())

Dirent(entryname = character(), entry, writable = FALSE)

requireManifest(inputID, sep = "\\\\n")

requireSubworkflow()

requireScatter()

requireMultipleInput()

requireStepInputExpression()
}
\arguments{
\item{docker}{The docker pull address.}

\item{Load}{dockerLoad}

\item{File}{dockerFile}

\item{Import}{dockerImport}

\item{ImageId}{dockerImageId}

\item{OutputDir}{dockerOutputDirectory}

\item{expressionLib}{optional code}

\item{packages}{The list of software to be configured.}

\item{listing}{The list of files or directories.}

\item{entryname}{The name of the file or subdirectory to create in
the output directory.}

\item{entry}{charactor or expression.}

\item{writable}{boolean.}

\item{inputID}{The input ID from corresponding `InputParam`.}

\item{sep}{The separator of the input files in the manifest config.}
}
\value{
A DockerRequirement list

A InlineJavascriptRequirement list

A SoftwareRequirement list

A InitialWorkDirRequirement list

A SubworkflowFeatureRequirement list

A ScatterFeatureRequirement list

A MultipleInputFeatureRequirement list

A StepInputExpressionRequirement list
}
\description{
requireDocker

requireJS

requireSoftware

InitialWorkDirRequirement

Dirent class

Create manifest for configure files

SubworkflowFeatureRequirement

ScatterFeatureRequirement

MultipleInputFeatureRequirement

StepInputExpressionRequirement
}
\details{
https://www.commonwl.org/v1.0/CommandLineTool.html#Dirent
}
\examples{
p1 <- InputParam(id = "ifiles", type = "File[]?", position = -1)
CAT = cwlProcess(baseCommand = "cat",
requirements = list(requireDocker("alpine"), requireManifest("ifiles"), requireJS()),
arguments = list("ifiles"),
inputs = InputParamList(p1))
}
