% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cwlParam.R
\name{runs}
\alias{runs}
\title{runs}
\usage{
runs(object)
}
\arguments{
\item{object}{A cwlStepParam object.}
}
\value{
cwlParam objects or paths of CWL file.
}
\description{
The function to access all runs of a cwlStepParam object
}
\examples{
s1 <- cwlStepParam()
runs(s1)
}
