% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cwlParam.R
\name{short}
\alias{short}
\title{short}
\usage{
short(object)
}
\arguments{
\item{object}{An cwlParam or cwlStepParam object}
}
\value{
A short summary of an object of cwlParam or cwlStepParam.
}
\description{
The function to show short summary of cwlParam or cwlStepParam
}
\examples{
s1 <- cwlStepParam()
short(s1)
}
