% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stepParam.R
\name{Step}
\alias{Step}
\title{Step function}
\usage{
Step(id, run = cwlParam(), In = list(), scatter = character(),
  scatterMethod = character())
}
\arguments{
\item{id}{The id of `stepParam` object.}

\item{run}{A `cwlParam` object for command tool.}

\item{In}{one or two layes of list.}

\item{scatter}{character or a list. The inputs to be scattered.}

\item{scatterMethod}{required if scatter is an array of more than
one element. It can be one of "dotproduct",
"nested_crossproduct" and "flat_crossproduct". Details:
https://www.commonwl.org/v1.0/Workflow.html#WorkflowStep}
}
\description{
Function to assign value to `stepParam` object.
}
