% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllClasses.R
\docType{class}
\name{stepInParamList-class}
\alias{stepInParamList-class}
\alias{stepInParamList}
\title{stepInParamList}
\usage{
stepInParamList(...)
}
\arguments{
\item{...}{A list of `stepInParam` objects.}
}
\value{
An object of class `stepInParamList`.
}
\description{
stepInParamList

stepInParamList
}
\examples{
s1 <- stepInParam(id = "s1")
stepInParamList(s1)
}
