% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllClasses.R
\docType{class}
\name{stepParam-class}
\alias{stepParam-class}
\alias{stepParam}
\title{stepParam}
\usage{
stepParam(
  id,
  run = cwlParam(),
  In = stepInParamList(),
  Out = list(),
  scatter = character(),
  scatterMethod = character(),
  label = character(),
  doc = character(),
  requirements = list(),
  hints = list(),
  when = character()
)
}
\arguments{
\item{id}{The unique identifier for this workflow step.}

\item{run}{A `cwlParam` object or the path of a cwl file.}

\item{In}{A `stepInParamList`.}

\item{Out}{A list of outputs}

\item{scatter}{character or a list. The inputs to be scattered.}

\item{scatterMethod}{required if scatter is an array of more than
one element. It can be one of "dotproduct",
"nested_crossproduct" and "flat_crossproduct". Details:
https://www.commonwl.org/v1.0/Workflow.html#WorkflowStep}

\item{label}{A short, human-readable label of this object.}

\item{doc}{A documentation string for this object, or an array of
strings which should be concatenated.}

\item{requirements}{requirements that apply to either the runtime
environment or the workflow engine.}

\item{hints}{hints applying to either the runtime environment or
the workflow engine}

\item{when}{If defined, only run the step when the expression
evaluates to true. If false the step is skipped.}
}
\value{
An object of class `stepParam`.
}
\description{
A workflow step parameters. More details:
https://www.commonwl.org/v1.0/Workflow.html#WorkflowStep
}
\examples{
s1 <- stepParam(id = "s1")
}
