% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllClasses.R
\docType{class}
\name{stepParam-class}
\alias{stepParam-class}
\alias{stepParam}
\title{stepParam}
\usage{
stepParam(id, run = cwlParam(), In = stepInParamList(), Out = list(),
  scatter = character(), scatterMethod = character())
}
\arguments{
\item{id}{The unique identifier for this workflow step.}

\item{run}{A `cwlParam` object.}

\item{In}{A `stepInParamList`.}

\item{Out}{A list of outputs}

\item{scatter}{character or a list. The inputs to be scattered.}

\item{scatterMethod}{required if scatter is an array of more than one element. It can be one of "dotproduct", "nested_crossproduct" and "flat_crossproduct". Details: https://www.commonwl.org/v1.0/Workflow.html#WorkflowStep}
}
\description{
stepParam

stepParam
A workflow step parameters. More details: https://www.commonwl.org/v1.0/Workflow.html#WorkflowStep
}
