% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/writeCWL.R
\name{writeCWL}
\alias{writeCWL}
\title{Write CWL
write `cwlParam` to cwl and yml.}
\usage{
writeCWL(cwl, prefix, ...)
}
\arguments{
\item{cwl}{A `cwlParam` or `cwlStepParam` object.}

\item{prefix}{The prefix of `cwl` and `yml` file to write.}

\item{...}{Other options from `yaml::write_yaml`.}
}
\description{
Write CWL
write `cwlParam` to cwl and yml.
}
