% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/writeCWL.R
\name{writeCWL}
\alias{writeCWL}
\title{Write CWL}
\usage{
writeCWL(cwl, prefix, noDocker = FALSE, ...)
}
\arguments{
\item{cwl}{A `cwlParam` or `cwlStepParam` object.}

\item{prefix}{The prefix of `cwl` and `yml` file to write.}

\item{noDocker}{Whether to disable docker.}

\item{...}{Other options from `yaml::write_yaml`.}
}
\description{
write `cwlParam` to cwl and yml.
}
